+++ readium/shared/r2-shared/src/main/java/org/readium/r2/shared/util/URITemplate.kt
@@ -59,6 +59,7 @@ data class URITemplate(val uri: String) {
 
         return Href(expanded).percentEncodedString
             .replace("~~%20~~", "%2B")
+            .replace("~~+~~", "%2B")
     }
 
 }