+++ readium/shared/r2-shared/src/test/java/org/readium/r2/shared/extensions/BitmapTest.kt
@@ -12,16 +12,14 @@ package org.readium.r2.shared.extensions
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.util.Size
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.robolectric.annotation.Config
+import org.robolectric.RobolectricTestRunner
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
 
-@RunWith(AndroidJUnit4::class)
-@Config(sdk = [28])
+@RunWith(RobolectricTestRunner::class)
 class BitmapTest {
 
     private val bitmap: Bitmap