+++ readium/shared/r2-shared/src/test/java/org/readium/r2/shared/fetcher/ArchiveFetcherTest.kt
@@ -10,7 +10,6 @@
 package org.readium.r2.shared.fetcher
 
 import android.webkit.MimeTypeMap
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import kotlinx.coroutines.runBlocking
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -19,15 +18,14 @@ import org.readium.r2.shared.linkBlocking
 import org.readium.r2.shared.publication.Link
 import org.readium.r2.shared.publication.Properties
 import org.readium.r2.shared.readBlocking
+import org.robolectric.RobolectricTestRunner
 import org.robolectric.Shadows
-import org.robolectric.annotation.Config
 import java.nio.charset.StandardCharsets
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertNotNull
 
-@RunWith(AndroidJUnit4::class)
-@Config(sdk = [28])
+@RunWith(RobolectricTestRunner::class)
 class ArchiveFetcherTest {
 
     private val fetcher: Fetcher