+++ readium/shared/r2-shared/src/test/java/org/readium/r2/shared/publication/services/CoverServiceTest.kt
@@ -12,7 +12,6 @@ package org.readium.r2.shared.publication.services
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.util.Size
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import kotlinx.coroutines.runBlocking
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -20,14 +19,13 @@ import org.readium.r2.shared.fetcher.FileFetcher
 import org.readium.r2.shared.linkBlocking
 import org.readium.r2.shared.publication.*
 import org.readium.r2.shared.readBlocking
-import org.robolectric.annotation.Config
+import org.robolectric.RobolectricTestRunner
 import java.io.File
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
 
-@RunWith(AndroidJUnit4::class)
-@Config(sdk = [28])
+@RunWith(RobolectricTestRunner::class)
 class CoverServiceTest {
 
     private val coverBytes: ByteArray