+++ readium/shared/r2-shared/src/test/java/org/readium/r2/shared/util/URITemplateTest.kt
@@ -10,8 +10,11 @@
 package org.readium.r2.shared.util
 
 import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
 import kotlin.test.assertEquals
 
+@RunWith(RobolectricTestRunner::class)
 class URITemplateTest {
 
     @Test